Remove Rest folder
