+++ src/Pim/Bundle/ApiBundle/Controller/CategoryController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\ApiBundle\Controller\Rest;
+namespace Pim\Bundle\ApiBundle\Controller;
 
 use Akeneo\Component\Classification\Repository\CategoryRepositoryInterface;
 use Akeneo\Component\StorageUtils\Exception\PropertyException;
