+++ src/Pim/Bundle/ApiBundle/Resources/config/controllers.yml
@@ -1,15 +1,15 @@
 parameters:
-    pim_api.controller.rest.category.class: Pim\Bundle\ApiBundle\Controller\Rest\CategoryController
-    pim_api.controller.rest.family.class: Pim\Bundle\ApiBundle\Controller\Rest\FamilyController
-    pim_api.controller.rest.attribute.class: Pim\Bundle\ApiBundle\Controller\Rest\AttributeController
-    pim_api.controller.rest.attribute_option.class: Pim\Bundle\ApiBundle\Controller\Rest\AttributeOptionController
+    pim_api.controller.category.class: Pim\Bundle\ApiBundle\Controller\CategoryController
+    pim_api.controller.family.class: Pim\Bundle\ApiBundle\Controller\FamilyController
+    pim_api.controller.attribute.class: Pim\Bundle\ApiBundle\Controller\AttributeController
+    pim_api.controller.attribute_option.class: Pim\Bundle\ApiBundle\Controller\AttributeOptionController
     pim_api.controller.channel.class: Pim\Bundle\ApiBundle\Controller\ChannelController
     pim_api.controller.token.class: Pim\Bundle\ApiBundle\Controller\TokenController
     pim_api.controller.product.class: Pim\Bundle\ApiBundle\Controller\ProductController
 
 services:
-    pim_api.controller.rest.category:
-        class: '%pim_api.controller.rest.category.class%'
+    pim_api.controller.category:
+        class: '%pim_api.controller.category.class%'
         arguments:
             - '@pim_catalog.repository.category'
             - '@pim_serializer'
@@ -20,8 +20,8 @@ services:
             - '@router'
             - 'https://docs.akeneo.com/%s/reference/standard_format/other_entities.html#category'
 
-    pim_api.controller.rest.family:
-        class: '%pim_api.controller.rest.family.class%'
+    pim_api.controller.family:
+        class: '%pim_api.controller.family.class%'
         arguments:
             - '@pim_catalog.repository.family'
             - '@pim_serializer'
@@ -32,8 +32,8 @@ services:
             - '@router'
             - 'https://docs.akeneo.com/%s/reference/standard_format/other_entities.html#family'
 
-    pim_api.controller.rest.attribute:
-        class: '%pim_api.controller.rest.attribute.class%'
+    pim_api.controller.attribute:
+        class: '%pim_api.controller.attribute.class%'
         arguments:
             - '@pim_catalog.repository.attribute'
             - '@pim_serializer'
@@ -44,8 +44,8 @@ services:
             - '@router'
             - 'https://docs.akeneo.com/%s/reference/standard_format/other_entities.html#attribute'
 
-    pim_api.controller.rest.attribute_option:
-        class: '%pim_api.controller.rest.attribute_option.class%'
+    pim_api.controller.attribute_option:
+        class: '%pim_api.controller.attribute_option.class%'
         arguments:
             - '@pim_catalog.repository.attribute'
             - '@pim_catalog.repository.attribute_option'
