+++ src/Pim/Bundle/ApiBundle/Resources/config/routing/category.yml
@@ -1,19 +1,19 @@
 pim_api_rest_category_list:
     path: /categories
-    defaults: { _controller: pim_api.controller.rest.category:listAction, _format: json }
+    defaults: { _controller: pim_api.controller.category:listAction, _format: json }
     methods: [GET]
 
 pim_api_rest_category_get:
     path: /categories/{code}
-    defaults: { _controller: pim_api.controller.rest.category:getAction, _format: json }
+    defaults: { _controller: pim_api.controller.category:getAction, _format: json }
     methods: [GET]
 
 pim_api_rest_controller_create:
     path: /categories
-    defaults: { _controller: pim_api.controller.rest.category:createAction, _format: json}
+    defaults: { _controller: pim_api.controller.category:createAction, _format: json}
     methods: [POST]
 
 pim_api_rest_controller_partial_update:
     path: /categories/{code}
-    defaults: { _controller: pim_api.controller.rest.category:partialUpdateAction, _format: json}
+    defaults: { _controller: pim_api.controller.category:partialUpdateAction, _format: json}
     methods: [PATCH]
