+++ src/Pim/Bundle/ApiBundle/Resources/config/routing/family.yml
@@ -1,19 +1,19 @@
 pim_api_rest_family_list:
     path: /families
-    defaults: { _controller: pim_api.controller.rest.family:listAction, _format: json }
+    defaults: { _controller: pim_api.controller.family:listAction, _format: json }
     methods: [GET]
 
 pim_api_rest_family_get:
     path: /families/{code}
-    defaults: { _controller: pim_api.controller.rest.family:getAction, _format: json }
+    defaults: { _controller: pim_api.controller.family:getAction, _format: json }
     methods: [GET]
 
 pim_api_rest_family_create:
     path: /families
-    defaults: { _controller: pim_api.controller.rest.family:createAction, _format: json }
+    defaults: { _controller: pim_api.controller.family:createAction, _format: json }
     methods: [POST]
 
 pim_api_rest_family_partial_update:
     path: /families/{code}
-    defaults: { _controller: pim_api.controller.rest.family:partialUpdateAction, _format: json }
+    defaults: { _controller: pim_api.controller.family:partialUpdateAction, _format: json }
     methods: [PATCH]
